requirements.txt
ccxt==4.4.46
discord.py==2.4.0
requests==2.24.0
pandas==1.5.3
mplfinance==0.12.10b0
python-dotenv==1.0.1